requirements.txt
langchain
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
pypdf
python-dotenv
langdetect